pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gene-circuit-design"
version = "0.1.0"
description = "Active learning for genetic circuit design"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Cell Ethology Lab", email = "jerry@westlake.edu.cn"},
]
keywords = ["active-learning", "dna-sequences", "circuit-design", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "botorch>=0.16.1",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "llvmlite>=0.37.0",
    "numpy>=1.21.0",
    "openpyxl>=3.1.5",
    "pandas>=1.3.0",
    "pyyaml>=6.0.2",
    "safetensors>=0.5.3",
    "scikit-learn>=1.6.0",
    "scipy>=1.7.0",
    "setuptools>=65.0.0",
    "skorch>=1.0.0",
    "submitit>=1.5.3",
    "torch>=2.5.1,<2.9",
    "tqdm<=4.66.5",
]

[project.urls]
Homepage = "https://github.com/cellethology/gene-circuit-design"
Repository = "https://github.com/cellethology/gene-circuit-design.git"
Issues = "https://github.com/cellethology/gene-circuit-design/issues"

[project.scripts]
run-active-learning = "run_experiments:main"

[tool.hatch.build.targets.wheel]
packages = ["utils"]

[tool.uv]
index-url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "sys_platform == 'linux'" }]

[tool.ruff]
line-length = 88
target-version = "py310"
fix = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "F841",  # Local variable is assigned to but never used
]

[tool.ruff.lint.per-file-ignores]
"test/*" = ["S101"]  # Allow assert statements in tests

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--cov=core",
    "--cov=job_sub",
    "--cov=utils",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
]

[dependency-groups]
dev = [
    "coverage>=7.6.1",
    "ipykernel>=6.29.5",
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "pandas>=2.0.3",
    "pre-commit>=3.5.0",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.13",
    "safetensors>=0.5.3",
    "scikit-learn>=1.3.2",
    "seaborn>=0.13.2"
]

[tool.coverage.run]
source = [
    "core",
    "job_sub",
    "utils",
]
branch = true

[tool.coverage.report]
show_missing = true
skip_empty = true
